+++ src/components/chatview-components/ChatRoom.vue
@@ -33,7 +33,7 @@
           <!-- 
           <DropdownMenu v-if="isActiveDropdown" style="width: 400px">
             <template #dropdown-item>
-              <BasicListElement
+              <BasicListItem
                 v-for="e in friends"
                 :key="e.id"
                 :id="e.id"
@@ -82,7 +82,7 @@ import SetChannelPasswordModal from '@/components/chatview-components/modals/Set
 import MessageList from '@/components/chatview-components/MessageList.vue';
 import ChatInputBox from '@/components/chatview-components/ChatInputBox.vue';
 // import DropdownMenu from '@/components/dropdown-component/DropdownMenu.vue';
-// import BasicListElement from '@/components/BasicListElement.vue';
+// import BasicListItem from '@/components/BasicListItem.vue';
 import type { IChatInfo } from '@/interfaces/ChatInfo.interface';
 import type { IUser } from '@/interfaces/User.interface';
